+++ CHANGELOG.md
@@ -7,13 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.25.0] - 2025-10-09
+## [0.25.0] - 2025-10-15
 
 ### Changed
 
 - Default parse for non 10x `spatialdata` iterates all spatial tables
 - Slide identifier and index in dataset strings
 - Toggle canvas compression format and level for sufficiently large ROIs
+- Expand accepted tiff file extensions
+- Use CLI array store type for .h5ad imports
+
+### Fixed
+
+- Include current session mask uploads when parsing `zarr`
 
 ## [0.24.0] - 2025-09-08